Modified model and viewmodel, still a ways to go

+++ Diagram_WinProg2016/Model/Diagram_WinProg2016.Model/IClass.cs
@@ -9,12 +9,12 @@ namespace Diagram_WinProg2016.Model
 	interface IClass
 	{
 		int ID { get; set; }
-		int X { get; set; }
-		int Y { get; set; }
-		int Height { get; set; }
-		int Width { get; set; }
-		int GridCenterX { get; set; }
-		int GridCenterY { get; set; }
+		double X { get; set; }
+		double Y { get; set; }
+		double Height { get; set; }
+		double Width { get; set; }
+		double GridCenterX { get; set; }
+		double GridCenterY { get; set; }
 		string ClassName { get; set; }
 		List<string> Fields { get; set; }
 		List<string> Methods { get; set; }
